Fix: Declare a single class per file
