+++ src/LaunchDarkly/FeatureFlag.php
@@ -223,197 +223,3 @@ class FeatureFlag
         return $this->_deleted;
     }
 }
-
-class EvalResult
-{
-    private $_variation = null;
-    private $_value = null;
-    /** @var array */
-    private $_prerequisiteEvents = [];
-
-    /**
-     * EvalResult constructor.
-     * @param null $value
-     * @param array $prerequisiteEvents
-     */
-    public function __construct($variation, $value, array $prerequisiteEvents)
-    {
-        $this->_variation = $variation;
-        $this->_value = $value;
-        $this->_prerequisiteEvents = $prerequisiteEvents;
-    }
-
-    /**
-     * @return int
-     */
-    public function getVariation()
-    {
-        return $this->_variation;
-    }
-
-    /**
-     * @return null
-     */
-    public function getValue()
-    {
-        return $this->_value;
-    }
-
-    /**
-     * @return array
-     */
-    public function getPrerequisiteEvents()
-    {
-        return $this->_prerequisiteEvents;
-    }
-}
-
-class WeightedVariation
-{
-    /** @var int */
-    private $_variation = null;
-    /** @var int */
-    private $_weight = null;
-
-    private function __construct($variation, $weight)
-    {
-        $this->_variation = $variation;
-        $this->_weight = $weight;
-    }
-
-    public static function getDecoder()
-    {
-        return function ($v) {
-            return new WeightedVariation($v['variation'], $v['weight']);
-        };
-    }
-
-    /**
-     * @return int
-     */
-    public function getVariation()
-    {
-        return $this->_variation;
-    }
-
-    /**
-     * @return int
-     */
-    public function getWeight()
-    {
-        return $this->_weight;
-    }
-}
-
-class Target
-{
-    /** @var string[] */
-    private $_values = array();
-    /** @var int */
-    private $_variation = null;
-
-    protected function __construct(array $values, $variation)
-    {
-        $this->_values = $values;
-        $this->_variation = $variation;
-    }
-
-    public static function getDecoder()
-    {
-        return function ($v) {
-            return new Target($v['values'], $v['variation']);
-        };
-    }
-
-    /**
-     * @return \string[]
-     */
-    public function getValues()
-    {
-        return $this->_values;
-    }
-
-    /**
-     * @return int
-     */
-    public function getVariation()
-    {
-        return $this->_variation;
-    }
-}
-
-class Prerequisite
-{
-    /** @var string */
-    private $_key = null;
-    /** @var int */
-    private $_variation = null;
-
-    protected function __construct($key, $variation)
-    {
-        $this->_key = $key;
-        $this->_variation = $variation;
-    }
-
-    public static function getDecoder()
-    {
-        return function ($v) {
-            return new Prerequisite($v['key'], $v['variation']);
-        };
-    }
-
-    /**
-     * @return string
-     */
-    public function getKey()
-    {
-        return $this->_key;
-    }
-
-    /**
-     * @return int
-     */
-    public function getVariation()
-    {
-        return $this->_variation;
-    }
-}
-
-class Rollout
-{
-    /** @var WeightedVariation[] */
-    private $_variations = array();
-    /** @var string */
-    private $_bucketBy = null;
-
-    protected function __construct(array $variations, $bucketBy)
-    {
-        $this->_variations = $variations;
-        $this->_bucketBy = $bucketBy;
-    }
-
-    public static function getDecoder()
-    {
-        return function ($v) {
-            return new Rollout(
-                array_map(WeightedVariation::getDecoder(), $v['variations']),
-                isset($v['bucketBy']) ? $v['bucketBy'] : null);
-        };
-    }
-
-    /**
-     * @return WeightedVariation[]
-     */
-    public function getVariations()
-    {
-        return $this->_variations;
-    }
-
-    /**
-     * @return string
-     */
-    public function getBucketBy()
-    {
-        return $this->_bucketBy;
-    }
-}
