+++ src/LaunchDarkly/InvalidSDKKeyException.php
@@ -1,0 +1,9 @@
+<?php
+namespace LaunchDarkly;
+
+/**
+ * Used internally.
+ */
+class InvalidSDKKeyException extends \Exception
+{
+}
